% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homogenization_QC.R
\name{homogenization_QC}
\alias{homogenization_QC}
\title{Processing summary and quality control checks for homogenized LTER
  Soil Organic Matter Working Group data}
\usage{
homogenization_QC(directoryName, temporaryDirectory)
}
\arguments{
\item{directoryName}{The quoted name of the Google Drive directory containing
data and a key translation file}

\item{temporaryDirectory}{The quoted path and name of a temporary directory
on the user's local computer for storing script output. The directory does
not have to exist. The directory should end with a forward slash (Mac,
Linux) or backward slash (Windows).}
}
\value{
A report in PDF format of processing summary, quality control checks,
  and generalized plots of homogenized data in a local directory identified
  by the user.
}
\description{
The function homogonization_QC imports provided and homogenized
  data, and the associated key translation file from a Google Directory that
  the user is able to access. The script compares the provided data to the
  homogenized data, and reports on processing output, select quality control
  checks, and provides generalized plots of independent versus dependent
  variables. Output is a PDF report.
}
\note{
data_homogonization relies on the helper functions detect_os and
  sheet_download.
}
\examples{
\dontrun{

 homogenization_QC(directoryName = 'CDR_E120',
                   temporaryDirectory = '~/Desktop/CRD_E120_output/')

}

}
