% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheet_download.R
\name{sheet_download}
\alias{sheet_download}
\title{Import a Google Sheet}
\usage{
sheet_download(fileName, skipRows, missingValueCode)
}
\arguments{
\item{fileName}{name of the Google Sheet to import}

\item{skipRows}{number of rows to skip when importing (optional; default:0)}

\item{missingValueCode}{string to denote missing values (optional;
default:'NA')}
}
\value{
R object of type tibble
}
\description{
The sheet_download function downloads the contents of a Google
Sheet into an R object with with optional parameters for rows to skip and a
missing value code upon import.
}
\note{
sheet_download is intended primarily as a helper function for batch
downloading Google Sheets via a apply or map call but can be used
independently.
}
